+++ packages/ad_sdk/test/ad_sdk_test.dart
@@ -1,4 +1,4 @@
-import 'package:ad_sdk/ad_sdk.dart';
+import 'package:applovin_admob_sdk/applovin_admob_sdk.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {